requirements.txt
requests==2.31.0
ratelimiter==1.2.0.post0
psycopg2-binary==2.9.7